+++ Support/String.cc
@@ -57,7 +57,8 @@ vector<string> fabrique::Split(const std::string& s, const std::string delim)
 	}
 
 	assert(lastDelimiter >= -1);
-	parts.push_back(s.substr(static_cast<size_t>(lastDelimiter + delim.length())));
+	parts.push_back(s.substr(static_cast<size_t>(
+		static_cast<unsigned long long>(lastDelimiter) + delim.length())));
 
 	return parts;
 }
